+++ app/(auth)/signup/page.tsx
@@ -1,7 +1,7 @@
 import { getServerSession } from "next-auth";
 import Signup from "../../../components/common/Signup";
 import { redirect } from "next/navigation";
-import { authOptions } from "../../api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/auth";
 
 export default async function SignupPage() {
   const session = await getServerSession(authOptions);
