+++ eslint.config.mjs
@@ -10,6 +10,9 @@ const compat = new FlatCompat({
 });
 
 const eslintConfig = [
+  {
+    ignores: ["lib/generated/**/*"],
+  },
   ...compat.extends("next/core-web-vitals", "next/typescript"),
 ];
