+++ lib/auth.ts
@@ -1,0 +1,94 @@
+import CredentialsProvider from "next-auth/providers/credentials";
+import GoogleProvider from "next-auth/providers/google";
+import bcrypt from "bcryptjs";
+import type { Session, User, Account } from "next-auth";
+import type { JWT } from "next-auth/jwt";
+import type { NextAuthOptions, SessionStrategy } from "next-auth";
+import prisma from "@/lib/prisma";
+
+export const authOptions: NextAuthOptions = {
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID!,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+    }),
+    CredentialsProvider({
+      name: "Credentials",
+      credentials: {
+        email: {
+          label: "Email",
+          type: "email",
+          placeholder: "test@example.com",
+        },
+        password: { label: "Password", type: "password" },
+      },
+      async authorize(credentials) {
+        if (!credentials?.email || !credentials?.password) return null;
+        try {
+          const user = await prisma.user.findUnique({
+            where: { email: credentials.email },
+          });
+
+          if (!user) return null;
+          const isPasswordValid = await bcrypt.compare(
+            credentials.password,
+            user.password
+          );
+          if (!isPasswordValid) return null;
+          const { password, ...userWithoutPassword } = user;
+          return userWithoutPassword;
+        } catch (error) {
+          console.error("Auth error:", error);
+          return null;
+        }
+      },
+    }),
+  ],
+  session: { strategy: "jwt" as SessionStrategy },
+  pages: { signIn: "/signin" },
+  callbacks: {
+    async signIn({ user, account }: { user: User; account: Account | null }) {
+      if (account?.provider === "google") {
+        try {
+          const existingUser = await prisma.user.findUnique({
+            where: { email: user.email! },
+          });
+          if (!existingUser) {
+            await prisma.user.create({
+              data: { email: user.email!, name: user.name!, password: "" },
+            });
+          }
+          return true;
+        } catch (error) {
+          console.error("Google signin error:", error);
+          return false;
+        }
+      }
+      return true;
+    },
+    async jwt({ token, user }: { token: JWT; user?: User }) {
+      if (user || token.email) {
+        const dbUser = await prisma.user.findUnique({
+          where: { email: user?.email ?? token.email ?? undefined },
+        });
+
+        if (dbUser) {
+          token.id = dbUser.id;
+          token.email = dbUser.email;
+          token.name = dbUser.name;
+          token.role = dbUser.role;
+        }
+      }
+      return token;
+    },
+    async session({ session, token }: { session: Session; token: JWT }) {
+      if (token && session.user) {
+        session.user.id = token.id as string;
+        session.user.email = token.email as string;
+        session.user.name = token.name as string;
+        (session.user as any).role = token.role;
+      }
+      return session;
+    },
+  },
+};
