+++ lib/service/auth.ts
@@ -1,4 +1,4 @@
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/auth";
 
 export const getAuthSession = () => getServerSession(authOptions);
